feat(twin-slack): warm canvases/topic/emoji and domain split
Implement ruled warm gaps (canvases CRUD, setTopic/setPurpose, emoji.list)
and carve SlackDomain into domain/* modules.
