+++ packages/twin-slack/src/domain/index.ts
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: Apache-2.0
+//
+// Domain barrel (F-679): re-exports only.
+export { SlackDomain } from "./slack-domain.js";
+export type { Actor, DeltaHook } from "./slack-domain.js";
