+++ packages/twin-slack/test/mcp-legacy.test.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, it } from "vitest";
 import { createSlackTwinApp } from "../src/twin.js";
 import { openSlackTwinDatabase } from "../src/db.js";
-import { SlackDomain } from "../src/domain.js";
+import { SlackDomain } from "../src/domain/index.js";
 import { defaultSeedState } from "../src/seed.js";
 import { toolDefinitions } from "../src/tools.js";
 import { signTestToken, TEST_SID, withAuth } from "./_authHelper.js";
