+++ packages/twin-slack/test/performance.test.ts
@@ -6,7 +6,7 @@
 
 import { beforeAll, describe, expect, it } from "vitest";
 import { openSlackTwinDatabase } from "../src/db.js";
-import { SlackDomain } from "../src/domain.js";
+import { SlackDomain } from "../src/domain/index.js";
 import { defaultSeedState } from "../src/seed.js";
 
 const SEED_MESSAGE_COUNT = 1000;
